backend/requirements.txt
fastapi
uvicorn
gunicorn
torch
torchvision
opencv-python-headless
python-dotenv
python-multipart
opencv-python-headless
urllib3
ffmpeg-python
thefuzz
pyenchant
wordcloud
pydantic
boto3
amazon-transcribe
pydantic
pydantic_settings